TextDomain: avoid `Options` that are never empty
These two fields of `TextDomain` *always* contain a value:

* `name` is initialized in `TextDomain::new`, and there is no way to
  change it afterwards;
* `locale_category` is also initialized in `TextDomain::new`, and even
  though there's a setter to change the field later, the setter doesn't
  allow storing a `None` in the field.

Theoretically, someone could default-construct an instance of
`TextDomain`, but it will be useless, as the only things one could do
with it are:

1. pass it around;
2. change its fields via setters;
3. call `init()` to actually execute some C calls.

But `init()` would try to unwrap the two fields and panic if they
contain `None`. Thus, default-constructed instances are useless, and can
be left out of the picture. Indeed, this commit removes the `Default`
instance altogether, disallowing this invalid state.

Technically this is a breaking change, though as explained earlier, no
well-formed, useful code should be affected.