Move gettext-rs into its own subdirectory
This enabled me to add separate READMEs for gettext-rs, gettext-sys, and
the repo. That, in turn, gave me a way to update the docs such that each
README only describes the stuff that's relevant to that particular
crate. Finally, this gave me an opportunity to include a note about
GNU LGPL.

Fixes #31.