Constrain input types to `Into<String>`
Previous commit requires the user to configure gettext such that it
converts results to UTF-8. However, when gettext can't find
a translation, is returns the input string **with no conversions**.
Therefore, the input should also be in UTF-8, which in Rust means it
should be of `String` type or something convertible to it.