Expose fallibility of `textdomain()`
Also, the return type is changed from `String` to `Vec<u8>` to indicate
that the returned value can be anything, really. In reality, the domain
name is a part of the filepath, which is not guaranteed to be UTF-8.

This is a breaking change since it changes the signature of
`textdomain()`, adds a new variant to `TextDomainError`, and removes
`PartialEq` implementation for the latter.