Discourage using `setlocale` to change gettext encoding
We used to suggest that users call either `bind_textdomain_codeset` or
`setlocale`, but this advice didn't make much sense: `setlocale` sets
not just the encoding, but also the language. If the user (i.e.
a programmer) knew the target language in advance, they could just write
all their strings in that language, and not bother with gettext.